+++ documentation/src/components/counting-number/index.tsx
@@ -3,33 +3,30 @@ import { useInView } from "framer-motion";
 import { useAnimatedCounter } from "../../hooks/use-animated-counter";
 
 type CountingNumberProps = {
+    className?: string;
     from?: number;
     to: number;
     duration?: number;
-    format?: boolean;
+    format?: (value: number) => string;
     once?: boolean;
 };
 
 export const CountingNumber: React.FC<CountingNumberProps> = ({
-    // from,
+    className,
+    from,
     to,
-    // duration,
     format,
-    // once,
+    duration = 1,
+    once = true,
 }) => {
-    // const ref = React.useRef<HTMLSpanElement>(null);
-    // const inView = useInView(ref, { once });
+    const ref = React.useRef<HTMLSpanElement>(null);
+    const inView = useInView(ref, { once });
 
-    // const counter = useAnimatedCounter(inView ? to : from ?? 0, from, duration);
-
-    // const rounded = Math.round(counter);
+    const counter = useAnimatedCounter(inView ? to : from ?? 0, from, duration);
 
     return (
-        <span
-        /* ref={ref} */
-        >
-            {/* {format ? rounded.toLocaleString("en-US") : rounded} */}
-            {format ? to?.toLocaleString?.("en-US") ?? to ?? 0 : to}
+        <span ref={ref} className={className}>
+            {format ? format(counter) : counter}
         </span>
     );
 };