+++ examples/app-crm-minimal/src/routes/companies/list/index.tsx
@@ -1,0 +1,6 @@
+import React, { PropsWithChildren } from "react";
+import { CompanyList } from "../components";
+
+export const CompanyListPage: React.FC<PropsWithChildren> = ({ children }) => {
+    return <CompanyList>{children}</CompanyList>;
+};