+++ examples/app-crm-minimal/src/routes/kanban/create/index.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+import { ProjectModalCreate } from "../components";
+
+export const KanbanCreatePage: React.FC = () => {
+    return <ProjectModalCreate />;
+};