+++ packages/devtools/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @refinedev/devtools
 
+## 1.1.25
+
+### Patch Changes
+
+-   Updated dependencies [[`97d5d9c98b28`](https://github.com/refinedev/refine/commit/97d5d9c98b28a1d69cada0a746202f82bec45622)]:
+    -   @refinedev/cli@2.16.17
+
+## 1.1.24
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @refinedev/devtools-server@1.1.21
+    -   @refinedev/cli@2.16.16
+
 ## 1.1.23
 
 ### Patch Changes